requirements.txt
# Core dependencies for Agent-First Context System
# Requires Python 3.9+ (for qdrant-client and other dependencies)
click>=8.1.0
yamale>=4.0.0
PyYAML>=6.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.0.0

# Phase 3a: Vector Database
qdrant-client>=1.11.0,<2.0.0  # 1.11+ for Python 3.8 compatibility
openai>=0.27.0
aiofiles>=23.0.0  # Async file operations

# Phase 3b: Graph Database
neo4j>=5.0.0

# Phase 4: Key-Value Store
redis>=5.0.0
duckdb>=0.10.0
pandas>=2.0.0
numpy>=1.24.0

# Development dependencies
black>=23.0.0
mypy>=1.0.0
types-PyYAML>=6.0.0
types-redis>=4.0.0

# Future phases (commented out for now)
# promptlayer>=0.5.0